Preserve backward compatible behaviour of objectGUID attribute (#267)
In this fix the backward compatible behaviour of objectGUID attribute in kernel 4.4.8 and before is preserved by introducing
a userstore property 'transformObjectGUIDToUUID'. When the property is set to false the behaviour will be similar to
kernel 4.4.8. The default value of the property is true, to make sure any code that depends on the UUID format of the objectGUID
introduced in Kernel 4.4.9 won't break.